cni: move iptables-blocking code to separate file
